fix: 수익률 계산 부분에서 발생하는 부동 소수점 오류 수정
- double에서 BigDecimal을 활용한 계산 방식으로 변경

+++ src/main/java/lotto/view/output/Output.java
@@ -5,6 +5,7 @@ import lotto.message.ErrorMessage;
 import lotto.message.GuideMessage;
 import lotto.message.NumberType;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class Output {
@@ -43,7 +44,7 @@ public class Output {
                 + GuideMessage.PRIZE_UNIT.getGuideMessage());
     }
 
-    public void printProfitRate(double profitRate) {
+    public void printProfitRate(BigDecimal profitRate) {
         System.out.println(GuideMessage.PROFIT_RATE_PREFIX_MESSAGE.getGuideMessage()
                 + profitRate
                 + GuideMessage.PROFIT_RATE_SUFFIX_MESSAGE.getGuideMessage());
